backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.6.1
python-dotenv==1.0.0
